% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinal.R
\name{cardinal}
\alias{cardinal}
\title{Cardinal temperatures outputs}
\usage{
cardinal(
  data = "template dataframe",
  dayBL = "Diurnal bottom left temperature",
  dayBR = "Diurnal bottom right temperature",
  dayTL = "Diurnal top left temperature",
  dayTR = "Diurnal top right temperature",
  nightBL = "Nocturnal bottom left temperature",
  nightBR = "Nocturnal bottom right temperature",
  nightTL = "Nocturnal top left temperature",
  nightTR = "Nocturnal top right temperature",
  petri = "Number of petri in a column or row",
  method = "tbd",
  adjust = "tbd",
  fs = "lower thershold of fluctuation values",
  fe = "higher threshold of fluctuation values"
)
}
\arguments{
\item{data}{A template dataframe}

\item{dayBL}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{dayBR}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{dayTL}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{dayTR}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{nightBL}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{nightBR}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{nightTL}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{nightTR}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{petri}{Number of Petri dishes in a column or a row.}

\item{method}{Leave blank to use average corner temperature values or use "precise" to create a temperature gradient based in individual corner temperatures}

\item{adjust}{tbd, but will be used to center temperature into Petri dish}

\item{fs}{Lower fluctuation threshold}

\item{fe}{Higher fluctuation thresholds}
}
\value{
Cardinal temperatures, sub and supraoptimal equations and a visualization
}
\description{
Calculates and plot cardinal temperatures and sub/supra regression equations after selecting a temperature fluctuation threshold
}
