% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coolbear.R
\name{coolbear}
\alias{coolbear}
\title{Obtain T50 values}
\usage{
coolbear(
  day = "vector containing scoring dats",
  cumulative = "vector containing cumulative germination for each day",
  n = "total number of viable seeds"
)
}
\arguments{
\item{day}{A vector containing scoring days}

\item{cumulative}{A vector containing the cumulative germination for each day}

\item{n}{Number of viable seeds in the Petri dish (Total sown seeds - empty seeds)}
}
\value{
t50 value in days
}
\description{
Obtain T50 values using Coolbear et al. (1984) formula modified by Farooq et al.(2005). Vector containing scoring days must have same length than vector containing cumulative germination data
}
\examples{
coolbear(seq(1,10,1),c(0,0,0,0,0,6,12,15,18,20),20)


}
