% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_results2.R
\name{grid_results2}
\alias{grid_results2}
\title{Generate grid with results}
\usage{
grid_results2(
  x = "Template with cumulative germination data",
  dayBL = "Diurnal bottom left temperature",
  dayBR = "Diurnal bottom right temperature",
  dayTL = "Diurnal top left temperature",
  dayTR = "Diurnal top right temperature",
  nightBL = "Nocturnal bottom left temperature",
  nightBR = "Nocturnal bottom right temperature",
  nightTL = "Nocturnal top left temperature",
  nightTR = "Nocturnal top right temperature",
  petri = "Number of petri in a column or row",
  method = "average corners temperature or use these independently",
  adjust = "adjust temperature to center of Petri dish"
)
}
\arguments{
\item{x}{germination data template}

\item{dayBL}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{dayBR}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{dayTL}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{dayTR}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{nightBL}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{nightBR}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{nightTL}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{nightTR}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{petri}{Number of Petri dishes in a column or a row.}

\item{method}{Leave blank to use average corner temperature values or use "precise" to create a temperature gradient based in individual corner temperatures}

\item{adjust}{use "adjust=TRUE" to adjust temperature to Petri dish center.#'}
}
\value{
A data frame with day and night temperatures, average temperature and temperature fluctuation for each Petri dish
}
\description{
Taking into account thermal gradient plate corner temperatures this function return a data frame with day and night temperatures for each petri dish, their average temperature, the temperature fluctuation or the final germination \%. Using the "precise" method returns a more accurate estimate of temperatures across the thermal plate, while with the adjust parameter is possible to estimate the temperature at the center of each Petri dish
}
\examples{
data<-tg_example
grid_results(data, 0,3,40,38,0,38,2,39, petri=13)


}
