% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempgrad.R
\name{tempgrad}
\alias{tempgrad}
\title{Generate temperature gradient grid}
\usage{
tempgrad(
  dayBL = "Diurnal bottom left temperature",
  dayBR = "Diurnal bottom right temperature",
  dayTL = "Diurnal top left temperature",
  dayTR = "Diurnal top right temperature",
  nightBL = "Nocturnal bottom left temperature",
  nightBR = "Nocturnal bottom right temperature",
  nightTL = "Nocturnal top left temperature",
  nightTR = "Nocturnal top right temperature",
  petri = "Number of petri in a column or row",
  method = "tbd",
  adjust = "tbd"
)
}
\arguments{
\item{dayBL}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{dayBR}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{dayTL}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{dayTR}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{nightBL}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{nightBR}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{nightTL}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{nightTR}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{petri}{Number of Petri dishes in a column or a row.}

\item{method}{Leave blank to use average corner temperature values or use "precise" to create a temperature gradient based in individual corner temperatures}

\item{adjust}{tbd, but will be used to center temperature into Petri dish}
}
\value{
A dataframe with day and night temperatures for each Petri dish
}
\description{
Taking into account thermal gradient plate corner temperatures this function return a data frame with the temperatures of each petri dish. Using the "precise" method returns a more accurate estimate of temperatures across the thermal plate, while with the adjust parameter is possible to estimate the temperature at the center of each Petri dish
}
\examples{
tempgrad(dayBL=0,dayBR=0,dayTL=40,dayTR=40,nightBL=0,nightBR=40,nightTL=0,nightTR=40,petri=13)

}
