% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_example.R
\name{tg_example}
\alias{tg_example}
\title{Load example}
\usage{
tg_example()
}
\value{
an example data frame
}
\description{
Loads a dataframe which can be used as an example to try out the functions
}
\examples{
df<-tg_example()

}
