% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t50_in_progress.R
\name{cardinal}
\alias{cardinal}
\title{Find and plot cardinal temperatures}
\usage{
cardinal(
  data = "template dataframe",
  fs = "lower thershold of fluctuation values",
  fe = "higher threshold of fluctuation values"
)
}
\arguments{
\item{data}{A template dataframe}

\item{fs}{Lower fluctuation threshold}

\item{fe}{Higher fluctuation thresholds}
}
\value{
A plot with cardinal temperatures
}
\description{
Find and plot cardinal temperatures after selecting a temperature fluctuation threshold
}
