% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results}
\alias{plot_results}
\title{Plots thermal gradient temperature for each Petri dish}
\usage{
plot_results(
  x = "Template with cumulative germination data",
  dayBL = "Diurnal bottom left temperature",
  dayBR = "Diurnal bottom right temperature",
  dayTL = "Diurnal top left temperature",
  dayTR = "Diurnal top right temperature",
  nightBL = "Nocturnal bottom left temperature",
  nightBR = "Nocturnal bottom right temperature",
  nightTL = "Nocturnal top left temperature",
  nightTR = "Nocturnal top right temperature",
  petri = "Number of petri in a column or row",
  method = "tbd",
  adjust = "tbd",
  toplot = "show average temperature or temperature fluctuation"
)
}
\arguments{
\item{x}{germination data template}

\item{dayBL}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{dayBR}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{dayTL}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{dayTR}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{nightBL}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{nightBR}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{nightTL}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{nightTR}{Average diurnal temperature at the bottom left side of the thermal gradient plate}

\item{petri}{Number of Petri dishes in a column or a row.}

\item{method}{Leave blank to use average corner temperature values or use "precise" to create a temperature gradient based in individual corner temperatures}

\item{adjust}{tbd, but will be used to center temperature into Petri dish}

\item{toplot}{Decides between plotting average Petri dish temperature (="average") or fluctuation (="fluctuation")}
}
\value{
A graph with average Petri dish temperature or temperature fluctuation
}
\description{
Similar to tempgrad(), this function outputs a plot representing average temperature or temperature fluctuation for each petri dish
}
