% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tempgrad.R
\name{plot_tempgrad}
\alias{plot_tempgrad}
\title{Plots thermal gradient temperature for each Petri dish}
\usage{
plot_tempgrad(
  x = "output dataframe from tempgrad function",
  toplot = "tbdescribed"
)
}
\arguments{
\item{x}{a dataframe outputted by tempgrad function}

\item{toplot}{Decides between plotting average Petri dish temperature (="average) or fluctuation (="fluctuation)}
}
\value{
A graph with average Petri dish temperature or temperature fluctuation
}
\description{
Taking into account thermal gradient plate corner temperatures this function return a graph with the temperatures of each Petri dish. Using the "precise" method returns a more accurate estimated grid of temperatures across the thermal plate, while with the adjust parameter is possible to estimate the temperature at the center of each Petri dish
}
\examples{
x<-tempgrad(dayBL=0,dayBR=0,dayTL=40,dayTR=40,nightBL=0,nightBR=40,nightTL=0,nightTR=40,petri=13)
plot_tempgrad(x,toplot="average")
plot_tempgrad(x,toplot="fluctuation")
x\%>\%plot_tempgrad(toplot="average")
#'
}
