% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tg_example}
\alias{tg_example}
\title{Thermal gradient plate}
\format{
A data frame with 53940 rows and 10 variables:
}
\usage{
tg_example
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
